WIP: use a function for generating packages
Additionally in this commit we use the scanner function to find files
that should be packaged.

We can't actually get any information out of the scanner. The scanner
as it existed really just did some wiring up of dependencies in SCons
which is fine and it still needs to do that. But we can also use it as
a function to retrieve the list of files that belong to a given
component. It isn't the fastest solution in the world but it's
functional. We need to make the scanner also extend the components and
roles of children it finds still so we can use that when de-duping in
"real" package builders.