WIP: Introduce the env var DEST_DIR
When dest-dir and prefix are different we need to know where dest-dir
really was because INSTALL_DIR will actually be the concatenation of
DEST_DIR + PREFIX. So to find the root of all the installed
files (e.g. for packaging purposes) we need to track where dest-dir
was set to.